Alright, I made quite a few changes. I made it so that when you run it, Character one acts as the first player and persists until killed. I added a next button that creates a new, random opponent (the only problem is that the image doesn't update...)
In addition, I added a text file with the level growth rates of a lot of (but not all of) the characterstypes. I also added a few more character types.

I also implemented leveling up \m/(^.^)\m/ So yeah.

I fixed a few things with how attacks and stuff were calculated, and I think that's about it.

I also added a handy constructor to Weapon that creates a weapon given it's name. That's all.